[feat]cookieの処理をHeaderに移行
